+++ src/test/java/be/ugent/rml/TestCore.java
@@ -3,6 +3,8 @@ package be.ugent.rml;
 import be.ugent.rml.functions.FunctionLoader;
 import be.ugent.rml.records.RecordsFactory;
 import be.ugent.rml.store.QuadStore;
+import be.ugent.rml.store.RDF4JStore;
+import be.ugent.rml.term.NamedNode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.eclipse.rdf4j.rio.RDFFormat;
@@ -27,7 +29,8 @@ abstract class TestCore {
         File mappingFile = new File(mapPath);
         QuadStore rmlStore = Utils.readTurtle(mappingFile);
 
-        return new Executor(rmlStore, new RecordsFactory(new DataFetcher(mappingFile.getParent(), rmlStore)));
+        return new Executor(rmlStore,
+                new RecordsFactory(new DataFetcher(mappingFile.getParent(), rmlStore)));
     }
 
     Executor createExecutor(String mapPath, FunctionLoader functionLoader) throws IOException {
@@ -36,7 +39,8 @@ abstract class TestCore {
         File mappingFile = new File(classLoader.getResource(mapPath).getFile());
         QuadStore rmlStore = Utils.readTurtle(mappingFile);
 
-        return new Executor(rmlStore, new RecordsFactory(new DataFetcher(mappingFile.getParent(), rmlStore)), functionLoader);
+        return new Executor(rmlStore, new RecordsFactory(new DataFetcher(mappingFile.getParent(), rmlStore)),
+                functionLoader);
     }
 
     Executor doMapping(String mapPath, String outPath) {
@@ -53,23 +57,10 @@ abstract class TestCore {
     }
 
     void doMapping(Executor executor, String outPath) {
-        ClassLoader classLoader = getClass().getClassLoader();
-
         try {
             QuadStore result = executor.execute(null);
             result.removeDuplicates();
-
-            // load output file
-            File outputFile = new File(classLoader.getResource(outPath).getFile());
-            QuadStore outputStore;
-
-            if (outPath.endsWith(".nq")) {
-                outputStore = Utils.readTurtle(outputFile, RDFFormat.NQUADS);
-            } else {
-                outputStore = Utils.readTurtle(outputFile);
-            }
-
-            assertEquals(outputStore.toSortedString(), result.toSortedString());
+            compareStores(result, filePathToStore(outPath));
         } catch (IOException e) {
             logger.error(e.getMessage(), e);
             fail();
@@ -87,7 +78,55 @@ abstract class TestCore {
             Executor executor = new Executor(rmlStore, new RecordsFactory(new DataFetcher(mappingFile.getParent(), rmlStore)));
             QuadStore result = executor.execute(null);
         } catch (IOException e) {
-
+            // I expected you!
         }
+    }
+
+    private void compareStores(QuadStore store1, QuadStore store2) {
+        String string1 = store1.toSortedString();
+        String string2 = store2.toSortedString();
+
+        assertEquals(string2, string1);
+    }
+
+    void compareFiles(String path1, String path2, boolean removeTimestamps) {
+        RDF4JStore store1 = filePathToStore(path1);
+        RDF4JStore store2 = filePathToStore(path2);
+
+        if (removeTimestamps) {
+            store1.removeQuads(null, new NamedNode("http://www.w3.org/ns/prov#generatedAtTime"), null);
+            store2.removeQuads(null, new NamedNode("http://www.w3.org/ns/prov#generatedAtTime"), null);
+            store1.removeQuads(null, new NamedNode("http://www.w3.org/ns/prov#endedAtTime"), null);
+            store2.removeQuads(null, new NamedNode("http://www.w3.org/ns/prov#endedAtTime"), null);
+            store1.removeQuads(null, new NamedNode("http://www.w3.org/ns/prov#startedAtTime"), null);
+            store2.removeQuads(null, new NamedNode("http://www.w3.org/ns/prov#startedAtTime"), null);
+        }
+
+        try {
+            assertEquals(store1, store2);
+        } catch (AssertionError e) {
+            compareStores(store1, store2);
+        }
+    }
+
+    private RDF4JStore filePathToStore(String path) {
+        // load output-turtle file
+        File outputFile = null;
+        try {
+            outputFile = Utils.getFile(path);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        RDF4JStore store;
+
+        if (path.endsWith(".nq")) {
+            store = Utils.readTurtle(outputFile, RDFFormat.NQUADS);
+        } else if (path.endsWith(".json")) {
+            store = Utils.readTurtle(outputFile, RDFFormat.JSONLD);
+        } else {
+            store = Utils.readTurtle(outputFile);
+        }
+
+        return store;
     }
 }