+++ CHANGELOG.md
@@ -7,10 +7,27 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
-- support for NoSQL databases
-- support for Web APIs
-- support for SPARQL endpoints
-- support for TPF servers
+## [4.1.0] - 2018-10-15
+
+### Added
+
+- output formats: turtle, trig, trix, jsonld
+- functions with list input
+- metadata generation
+
+### Changed
+
+- use environment variable for RDB testing
+- all (including templating) works with function generators
+
+### Fixed
+
+- double join condition
+- support quote in literal
+- function file path not passed on to the function loader when using the CLI
+- when parsing a string into an RDF term, take into account @ is not always a language tag
+- parenttriplesmap handling null values
+- config file
 
 ## [4.0.0] - 2018-09-03
 
@@ -80,6 +97,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - support for accessing remote files (via HTTP GET)
 - basic support for functions
 
+[4.1.0]: https://github.com/RMLio/rmlmapper-java/compare/v4.0.0...v4.1.0
 [4.0.0]: https://github.com/RMLio/rmlmapper-java/compare/v0.2.1...v4.0.0
 [0.2.1]: https://github.com/RMLio/rmlmapper-java/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/RMLio/rmlmapper-java/compare/v0.1.6...v0.2.0